feat: DashboardModal이 정상적으로 대시보드를 출력하도록 개선
